fix architecture(to MVC)
